+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * `user` `groups` correction until Ansible 2.3
 * `gem` module now works and does not conflict with the Ruby keyword
 * `ec2_group` validation fixes for `vpc_id`, `purge_rules`, `purge_rules_egress`
+* `lineinfile` module - added missing `validate` parameter
 
 # 1.0.6 - 14 October 2016
 * Allow sharing variables between tasks (#29)